pyproject.toml
[tool.poetry]
name = "fast-socket"
version = "0.1.0"
description = ""
authors = ["TEMHITHORPHE <tehmhithorphe@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
socketify = "^0.0.30"
msgspec = "^0.18.6"
websocket-client = "^1.8.0"

[tool.virtua]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
